+++ R/methods_bayesQR.R
@@ -11,7 +11,7 @@ model_parameters.bayesQR <- function(model,
                                      diagnostic = c("ESS", "Rhat"),
                                      priors = TRUE,
                                      keep = NULL,
-                                     omit = NULL,
+                                     drop = NULL,
                                      parameters = keep,
                                      verbose = TRUE,
                                      ...) {
@@ -31,7 +31,7 @@ model_parameters.bayesQR <- function(model,
     diagnostic = diagnostic,
     priors = priors,
     keep_parameters = keep,
-    omit_parameters = omit,
+    drop_parameters = drop,
     verbose = verbose,
     ...
   )