+++ R/methods_brms.R
@@ -18,7 +18,7 @@ model_parameters.brmsfit <- function(model,
                                      standardize = NULL,
                                      group_level = FALSE,
                                      keep = NULL,
-                                     omit = NULL,
+                                     drop = NULL,
                                      parameters = keep,
                                      verbose = TRUE,
                                      ...) {
@@ -62,7 +62,7 @@ model_parameters.brmsfit <- function(model,
       component = component,
       standardize = standardize,
       keep_parameters = keep,
-      omit_parameters = omit,
+      drop_parameters = drop,
       verbose = verbose,
       ...
     )