+++ R/methods_cplm.R
@@ -37,7 +37,7 @@ model_parameters.zcpglm <- function(model,
                                     robust = FALSE,
                                     p_adjust = NULL,
                                     keep = NULL,
-                                    omit = NULL,
+                                    drop = NULL,
                                     parameters = keep,
                                     verbose = TRUE,
                                     ...) {
@@ -61,7 +61,7 @@ model_parameters.zcpglm <- function(model,
       robust = robust,
       p_adjust = p_adjust,
       keep_parameters = keep,
-      omit_parameters = omit,
+      drop_parameters = drop,
       ...
     )
   }