+++ R/methods_plm.R
@@ -57,7 +57,7 @@ model_parameters.pgmm <- function(model,
                                   robust = TRUE,
                                   p_adjust = NULL,
                                   keep = NULL,
-                                  omit = NULL,
+                                  drop = NULL,
                                   parameters = keep,
                                   verbose = TRUE,
                                   ...) {
@@ -71,7 +71,7 @@ model_parameters.pgmm <- function(model,
     robust = robust,
     p_adjust = p_adjust,
     keep_parameters = keep,
-    omit_parameters = omit,
+    drop_parameters = drop,
     ...
   )